PXE boot fails due to missing read request for grub.cfg
The commit 16ee58ceb40b ("Fix 'BootHole' and related vulnerabilities")
includes an upstream commit 80e584d2d9a2 ("tftp: Do not use priority
queue") which is triggering the issue. As the latter does not address
a CVE and is not included in Buster grub2, an option is to revert it.
The solution taken though is to add the upstream commit b3f87350160e
("tftp: Roll-over block counter to prevent data packets timeouts") to
address the regression.

VRVDR-53776