PXE boot fails due to missing read request for grub.cfg
PXE boot fails despite initial tests passing by including a fix for a
regression caused by 80e584d2d9a2 ("tftp: Do not use priority queue").
Address this by not using this fix to simplify the tftp code path (so
the regression fix is not needed either). This is acceptable as the fix
does not address any CVEs.

VRVDR-53776